+++ omniva-tarptautines-woo.php
@@ -27,6 +27,7 @@ if (!defined('ABSPATH')) {
 
 define('OMNIVA_GLOBAL_VERSION', '1.0.12');
 define('OMNIVA_GLOBAL_BASENAME', plugin_basename(__FILE__));
+define('OMNIVA_GLOBAL_PLUGIN_DIR', plugin_dir_path( __FILE__ ));
 
 register_activation_hook(__FILE__, array( 'OmnivaTarptautinesWoo\Main', 'activated' ) );
 register_deactivation_hook( __FILE__, array( 'OmnivaTarptautinesWoo\Main', 'deactivated' ) );