+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 ### Added
 - Added the option to select the price of the carrier to get with or without VAT
 - Added additional service: Insurance
+- Mass action buttons in Omniva orders page
 
 ### Fixed
-- Fixed carrier name showing in manifest table
+- Fixed carrier name showing in Omniva orders table
 
 ## [1.0.12]
 ### Updated